requirements.txt
python-telegram-bot==20.7
PyPDF2==3.0.1
pikepdf==8.10.1
pdf2docx==0.5.6
pdfplumber==0.10.3
pdf2image==1.16.3
PyMuPDF==1.23.8
Pillow==10.1.0
python-dotenv==1.0.0